Show only questions with answers
On homepage, show only questions with answers.

Check in the questions object, if an answer exists for a particular question. If the answer does exist, go through the answers object to get the body of the corresponding answer. After extracting the questions and their corresponding answers, add both questions and answers to an array which can be iterated through in the view.