Added Facebook sign in
Added Facebook authentication for sign-up and sign in. When user clicks the sign in button on the nav, they will be led to a modal which shows the Facebook sign in button. Currently, only Facebook sign in is offered. (Perhaps later, offer Google or manual email address/password sign in).

The modal is a Bootstrap component. Facebook button and icon are from Font Awesome and Bootstrap Social libraries. Facebook authentication function is via Firebase.

When clicking on the Facbeook button, the loginFacebook controller is called which triggers facebook authentication method through the LoginFactory.