Get user's authentication status to determine sign in
Get user's authentication from Firebase to determine whether the user has already signed in or is signing in for the first time.

If the user has never signed in before (or if they have signed in at least once before but is currently signed out [which will be added later]), they will get a pop-up that asks them to sign in through Facebook.

If the use has already signed in, they should be able to see their avatar.