Added ui-router library to make multiple views per page easier
Added ui-router library in the config file (app.js) so that  we can have multiple views per page. With the built-in ngRoute that comes with Angular, it's difficult to have multiple views per page because each URL is associated with a distinct view and controller. That means that if we have one URL but wanted to break it up into partial views, it becomes hard.

However, as in the case of the homepage where we are using multiple partial views for the page (the header, the main content and the footer), it's easier to use the ui-router library. We can set a controller for each partial view and set the templateURL path. By doing this, we don't need to use the Angular approach of ng-include and use a hack to wrap the partial view html with a controller attribute.

See angular 1 ui-router documentation for more info: https://ui-router.github.io/ng1/