Refactored how to show signed in user's questions and answers
Refactored how to show signed in user's questions and answers.

Previously, I created a separate function for showing signed in user's questions and answers; this was unnecessary extra code since we are already calling getQuestions from the q&a factory and iterating through questions and answers when we execute getQuestionsWithAnswers for community questions and answers.

Instead, it's more efficient just to check whether user is signed in with an if/else statement when we execute getQuestionWithAnswers function for getting all questions and answers.